add search side bar
