+++ src/hooks/useFeedGenerator.tsx
@@ -2,15 +2,12 @@ import _ from "lodash";
 import { useCallback } from "react";
 import { useStore } from "@/stores";
 import { AppBskyFeedDefs, AppBskyActorDefs } from "@atproto/api";
-// TODO feedのpinのtoggleがサイドパネルで更新されてない
 
 export const useFeedGenerator = () => {
   const like = useStore((state) => state.like);
   const deleteLike = useStore((state) => state.deleteLike);
   const updatePreferences = useStore((state) => state.updatePreferences);
   const updateFeedViewer = useStore((state) => state.updateFeedViewer);
-  const updateSavedFeedViewer = useStore((state) => state.updateSavedFeedViewer);
-  const updatePinnedFeedViewer = useStore((state) => state.updatePinnedFeedViewer);
 
   const updateViewer = useCallback(
     (f: AppBskyFeedDefs.GeneratorView, resourceURI?: string) => {
@@ -52,9 +49,8 @@ export const useFeedGenerator = () => {
         return p;
       });
       updatePreferences(update);
-      updateSavedFeedViewer(feed);
     },
-    [updatePreferences, updateSavedFeedViewer]
+    [updatePreferences]
   );
 
   const onTogglePin = useCallback(
@@ -68,9 +64,8 @@ export const useFeedGenerator = () => {
         return p;
       });
       updatePreferences(update);
-      updatePinnedFeedViewer(feed);
     },
-    [updatePreferences, updatePinnedFeedViewer]
+    [updatePreferences]
   );
 
   const onLike = useCallback(
