+++ src/hooks/usePreference.tsx
@@ -3,15 +3,14 @@ import { useStore } from "@/stores";
 
 export const usePreference = () => {
   const preferences = useStore((state) => state.preferences);
-  const savedFeeds = useStore((state) => state.savedFeeds);
-  const pinnedFeeds = useStore((state) => state.pinnedFeeds);
   const getPreferences = useStore((state) => state.getPreferences);
+  const savedFeeds = useStore((state) => state.savedFeeds);
 
   if (_.isEmpty(preferences)) {
     throw getPreferences();
   }
 
-  return { savedFeeds, pinnedFeeds } as const;
+  return { preferences, savedFeeds } as const;
 };
 
 export default usePreference;
