+++ src/templates/HistoryLayout.tsx
@@ -1,11 +1,9 @@
-import { ReactNode, useCallback, useState } from "react";
-import { useNavigate, useLocation } from "react-router-dom";
+import { ReactNode } from "react";
+import { useNavigate } from "react-router-dom";
 import Stack from "@mui/material/Stack";
-import TextField from "@mui/material/TextField";
-import InputAdornment from "@mui/material/InputAdornment";
 import IconButton from "@mui/material/IconButton";
-import SearchRoundedIcon from "@mui/icons-material/SearchRounded";
 import ArrowBackIosNewRoundedIcon from "@mui/icons-material/ArrowBackIosNewRounded";
+import Search from "@/components/Search";
 
 type Props = {
   children: ReactNode;
@@ -16,29 +14,6 @@ type Props = {
 // TODO コンテンツの上じゃなくて横のほうがいいかもしれない
 export const HistoryLayout = (props: Props) => {
   const navigate = useNavigate();
-  const location = useLocation();
-  const [keyword, setKeyword] = useState<string>(props.keyword || "");
-  const [isFocus, setFocus] = useState<boolean>(false);
-
-  const onChangeSearch = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      setKeyword(e.currentTarget.value);
-    },
-    [setKeyword]
-  );
-
-  const onSearch = useCallback(
-    (e: React.KeyboardEvent<HTMLInputElement>) => {
-      if (e.key === "Enter") {
-        navigate(`${location.pathname}?q=${keyword}`);
-      }
-    },
-    [keyword, navigate, location]
-  );
-
-  const onFocus = useCallback(() => {
-    setFocus(!isFocus);
-  }, [isFocus, setFocus]);
 
   return (
     <>
@@ -46,27 +21,7 @@ export const HistoryLayout = (props: Props) => {
         <IconButton size="small" color="primary" onClick={() => navigate(-1)}>
           <ArrowBackIosNewRoundedIcon fontSize="inherit" />
         </IconButton>
-        {props.search && (
-          <TextField
-            fullWidth
-            autoComplete="off"
-            placeholder="search"
-            size="small"
-            value={keyword}
-            onChange={onChangeSearch}
-            onFocus={onFocus}
-            onBlur={onFocus}
-            onKeyDown={onSearch}
-            InputProps={{
-              sx: { borderRadius: 10 },
-              startAdornment: (
-                <InputAdornment position="start">
-                  <SearchRoundedIcon color={isFocus ? "primary" : "disabled"} />
-                </InputAdornment>
-              ),
-            }}
-          />
-        )}
+        {props.search && <Search keyword={props.keyword} />}
       </Stack>
       {props.children}
     </>
